rename and organize catalogs and files, add layouts and pages, setup HeaderPage
