+++ log_app/src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./lauouts/App";
+import App from "./layouts/App";
 import * as serviceWorker from "./serviceWorker";
 
 ReactDOM.render(
