+++ log_app/src/pages/ContactPage.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ContactPage = (props) => {
+  return <div>Kontakt</div>;
+};
+
+export default ContactPage;
