+++ log_app/src/pages/ErrorPage.jsx
