+++ log_app/src/pages/HelpPage.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const HelpPage = (props) => {
+  return <div>Pomoc</div>;
+};
+
+export default HelpPage;
