+++ log_app/src/pages/RegistrationPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 const RegistrationPage = () => {
-  return <h1>Rejrstracja</h1>;
+  return <h1>Rejestracja</h1>;
 };
 
 export default RegistrationPage;
