+++ log_app/src/scss/_header.scss
@@ -1,0 +1,27 @@
+.header {
+  display: flex;
+  flex-wrap: nowrap;
+  justify-content: space-between;
+  padding: 10px;
+  &__logo {
+    flex-basis: 50%;
+    height: 100px;
+    background-image: url(../images/pallets_640.jpg);
+    background-size: cover;
+  }
+  &__info {
+    flex-basis: 40%;
+  }
+}
+.info {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  button {
+    flex-basis: 30%;
+    background-color: $basic-color;
+    color: $text-first-color;
+    font-size: $s-font-size;
+    border: none;
+  }
+}
