+++ log_app/src/scss/_variables.scss
@@ -1,0 +1,12 @@
+$basic-color: #1a2f5e;
+$first-color: #3b0918;
+$second-color: #b8390e;
+$third-color: #d9ab30;
+
+$text-first-color: #fff;
+$text-second-color: #000;
+
+$s-font-size: 14px;
+$m-font-size: 18px;
+$l-font-size: 20px;
+$x-font-size: 24px;
