+++ log_app/src/scss/style.scss
@@ -1,3 +1,4 @@
 @import "./reset";
 @import "./variables";
 @import "./homePage";
+@import "./header";
